fix(nx-serverless): fix deploy test to mock fs-extra and call getArgs properly
